Implement policy enforcement by hooking libc.
Since maps and profiles are not portable at this time, we've changed the name of profiles to be the base64 encoding of the main executable's full path. This will change if we make profiles portable in the future. When no profile is available for a program, our hooks revert to libc's original behavior (i.e., immediately freeing chunks for reallocation, making UAF exploitation possible).