Performance optimizations.
Lifting to IRSBs was accounting for about 66% of our execution time while parsing PT traces, which is a waste of time because the only thing the IRSB was then used for was to check if the block ends with a call, which can be determined without a lift. In short, we completely removed pyvex from the analysis.