Bugfix: Prevent recursion when calculating CFGNode hashes.
We were calculating the hash of a CFGNode on-demand, which due to context can require hashing other CFGNodes, causing a recursion that could exceed Python's max depth. Since CFGNodes aren't modfied after initialization, we can calculate the hash at this point and store it for all future references, saving a lot of time.