Analysis redesign (#2)
This update heavily revamps the analysis.

Execution units are now discovered by first finding the heads of loops in the program and then merging them until only the outer-most loops remain. The intuition is that completing one iteration of any of these loops correlates to the completion of one "task" in the program. This is significantly faster than analyzing simple circuits, easier to manually verify the correctness of, and more closely captures the spirit of prior execution unit partitioning designs.

To find release points, rather than searching for frees at the tail-end of the units, we now search for allocations at the beginning. This allows for a simpler forward analysis that scales better to large programs.

The hooks have also been updated with minor optimizations and to work with the new profiles.